+++ CPT202Project/src/main/java/com/xjtlu/cpt202/cpt202Project/service/CommentService.java
@@ -6,7 +6,7 @@ import com.xjtlu.cpt202.cpt202Project.entity.Comment;
 import java.util.List;
 
 public interface CommentService {
-    public List<Comment> listCommentByBlogId(Long BlogId);
+    public List<Comment> listCommentByBlogId(int BlogId);
     public int addComment(Comment comment);
     public int deleteComment(Comment comment);
